backend/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
numpy==1.24.3
python-dotenv==1.0.0
google-generativeai>=0.3.0
